backend/requirements.txt
anyio>=4.6.2.post1
asyncio>=3.4.3
fastapi[standard]>=0.115.6
openai>=1.55.3
python-dotenv>=1.0.1
requests>=2.32.3
rich>=13.9.4
uvicorn>=0.32.1
pyinstaller>=6.11.1
mcp
pyyaml
sqlalchemy
aiosqlite
greenlet
alembic